+++ sakura_core/_os/Clipboard.cpp
@@ -368,9 +368,8 @@ bool Clipboard::GetText(NativeW* pMemBuf, bool* pbColumnSelect, bool* pbLineSele
 		HDROP hDrop = (HDROP)::GetClipboardData(CF_HDROP);
 		if (hDrop) {
 			TCHAR sTmpPath[_MAX_PATH + 1] = {0};
-			const int nMaxCnt = DragQueryFile(hDrop, 0xFFFFFFFF, NULL, 0);
-
-			for (int nLoop=0; nLoop<nMaxCnt; ++nLoop) {
+			const size_t nMaxCnt = DragQueryFile(hDrop, 0xFFFFFFFF, NULL, 0);
+			for (size_t nLoop=0; nLoop<nMaxCnt; ++nLoop) {
 				DragQueryFile(hDrop, nLoop, sTmpPath, _countof(sTmpPath) - 1);
 				// 2012.10.05 Moca ANSI版に合わせて最終行にも改行コードをつける
 				pMemBuf->AppendStringT(sTmpPath);
