+++ sakura_core/convert/Convert_ToZenhira.cpp
@@ -13,7 +13,7 @@ bool Converter_ToZenhira::DoConvert(NativeW* pData)
 	// 文字数が減ることはあっても増えることは無いので、これでＯＫ
 	std::vector<wchar_t> buf(pData->GetStringLength() + 1);
 	wchar_t* pBuf = &buf[0];
-	int nBufLen = 0;
+	size_t nBufLen = 0;
 	Convert_HankataToZenkata(pData->GetStringPtr(), pData->GetStringLength(), pBuf, &nBufLen);
 
 	// 全カナ→全角ひらがな
